Extract interfaces from TabView and TabViewItem (#121)
Refactored the TabView and TabViewItem types to be abstracted by interfaces with the intention of better support for use with DI and extensibility concerns. This will allow the `TabViewItem.TabViewItemIndicator` to be created via a factory pattern, which opens the possibility for a developer to roll their own indicator(s).